requirements.txt
# LLM & Embeddings
google-generativeai==0.8.3
sentence-transformers==3.3.1
transformers>=4.34.0

# Vector Database
chromadb==0.5.23

# API & HTTP
requests==2.32.3
httpx==0.28.1

# Re-ranking
cross-encoder==1.2.0

# Streamlit
streamlit==1.41.0

# Data Processing
pandas==2.2.3
numpy==2.2.1
pydantic==2.10.4
pydantic-settings==2.7.0

# NLP
spacy==3.8.3

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0
cachetools==5.5.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0
